Major fixes and upgrades
New exported files and plots after
learning is completed. For each
training session, those files
are saved in the runs/ folder.

Values of azimuth are now rounded
to a more premeditated quantity of
decimal places.
State space dimensions need to be augmented depending on the
resolution. The bigger the resolution, the bigger
needs to be the initial dimensions of the state space.

Reward function was minorly restructured.

The step functions were restructured/corrected.
The empty environment is working for both simple
and complex action spaces. There is still
working in progress for the obstacle environment case.

Some adjustments in the qlearning functions were made.

The main run_rl2d file was adjusted to receive one more arg.
Also some handling was designed to ensure bounded environment
dimensions.