Major changes
Reward function works finally: The max angle variation is atan2(-1,2)
because of the major variation that the new kinematics causes with v = 1.
Also the obstacle avoidance reward works. It penalizes strongly
if in the direction of obstacles. As smaller is the distances, the more
it penalizes. Like a potential field.
Now the learning is handling terminal states.
Line of sight formulation changed, and the "-" configuration format
is working the way it is supposed to.
Centered trapezoidal model kinematics is confirmed.
It is the one we want to work.
The agent is being initialized now with random orientations.
Also minor changes and debugging not worth to mention.