Graphics class is now reseting the environment
Some general cleaning and changes of names.

Changed reward and steps statistics to be plotted
as moving averages of epochs
instead of for each timestep.

Way of building discrete timestep is a bit better.

Reward function is updated.
It is working perfectly for the empty env.
Still not on point for the obst env.